Fixed welcome page tab navigation a11y issue.
