+++ packages/app/main/src/appMenuBuilder.ts
@@ -31,7 +31,7 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import { BotInfo, SharedConstants, UpdateStatus } from '@bfemulator/app-shared';
+import { BotInfo, isMac, SharedConstants, UpdateStatus } from '@bfemulator/app-shared';
 import { CommandServiceImpl, CommandServiceInstance, ConversationService } from '@bfemulator/sdk-shared';
 import { app, clipboard, Menu, MenuItem, MenuItemConstructorOptions, shell } from 'electron';
 
@@ -40,7 +40,6 @@ import { BotHelpers } from './botHelpers';
 import { Emulator } from './emulator';
 import { rememberTheme } from './state/actions/windowStateActions';
 import { TelemetryService } from './telemetry';
-import { isMac } from './utils';
 import { store } from './state';
 import { getLocalhostServiceUrl } from './utils/getLocalhostServiceUrl';
 import { getCurrentConversationId } from './state/helpers/chatHelpers';
